chore: WatchListController 병합
- tab bar controller 로직 리팩토링
- temp 파일 제거
