+++ ios/Daangn/Daangn/Global/TabBarController.swift
@@ -15,39 +15,31 @@ final class TabBarController: UITabBarController {
     }
     
     private func setViewControllers() {
+        let homeNavigationController = createNavigationController(with: HomeViewController(), title: "홈화면", tabBarIconName: "house")
+        let salesNavigationController = createNavigationController(with: SalesViewController(), title: "판매내역", tabBarIconName: "newspaper")
+        let watchNavigationController = createNavigationController(with: WatchlistViewController(), title: "관심목록", tabBarIconName: "heart")
+        let chatNavigationController = createNavigationController(with: ChatViewController(), title: "채팅", tabBarIconName: "message")
         
-        
-        let homeViewController = HomeViewController()
-        let homeNavigationController = UINavigationController(rootViewController: homeViewController)
-        homeNavigationController.navigationBar.prefersLargeTitles = false
-        homeNavigationController.tabBarItem.title = "홈화면"
-        homeNavigationController.tabBarItem.image = generateStyledImage(systemName: "house")
-        
-        let salesViewController = SalesViewController()
-        let salesNavigationController = UINavigationController(rootViewController: salesViewController)
-        salesNavigationController.navigationBar.prefersLargeTitles = false
-        salesNavigationController.tabBarItem.title = "판매내역"
-        salesNavigationController.tabBarItem.image = generateStyledImage(systemName: "newspaper")
-
-        let sellHistory = SellHistoryViewController()
-        let watchlist = WatchlistViewController()
-        let chat = ChatViewController()
-        let myAccount = MyAccountViewController()
-        // navbar를 비활성화하면 <back 버튼도 안생김
-        
-        watchlist.tabBarItem = UITabBarItem(title: "관심목록", image: UIImage(systemName: "heart")?.withTintColor(.black), selectedImage: nil)
-        chat.tabBarItem = UITabBarItem(title: "채팅", image: UIImage(systemName: "message")?.withTintColor(.black), selectedImage: nil)
-        myAccount.tabBarItem = UITabBarItem(title: "내 계정", image: UIImage(systemName: "person")?.withTintColor(.black), selectedImage: nil)
+        let myAccountViewController = MyAccountViewController()
+        myAccountViewController.tabBarItem.title = "내 계정"
+        myAccountViewController.tabBarItem.image = generateStyledImage(systemName: "person")
         
         let controllers: [UIViewController] = [
             homeNavigationController,
             salesNavigationController,
-            watchlist,
-            chat,
-            myAccount,
+            watchNavigationController,
+            chatNavigationController,
+            myAccountViewController,
         ]
-        
         setViewControllers(controllers, animated: false)
+    }
+    
+    private func createNavigationController(with viewController: UIViewController, title: String, tabBarIconName: String) -> UINavigationController {
+        let navigationController = UINavigationController(rootViewController: viewController)
+        navigationController.navigationBar.prefersLargeTitles = false
+        navigationController.tabBarItem.title = title
+        navigationController.tabBarItem.image = generateStyledImage(systemName: tabBarIconName)
+        return navigationController
     }
     
     private func generateStyledImage(systemName name: String) -> UIImage? {
